requirements.txt
ccxt
nest_asyncio
python-telegram-bot
pandas
google-generativeai